requirements.txt
numpy
pandas
matplotlib
jupyter
tweepy
instaloader
facebook-scraper
PyYAML
beautifulsoup4
wordcloud
spacy
trafilatura
requests
sentence-transformers
scikit-learn
gensim
rake-nltk
seaborn
tensorflow_hub
tensorflow
scipy